moved components again
